Created users app and the User model.
User model inherits from CrideModel which is basically the parent model
of all in this django project, and from Abstract User.
This model add some fields like the email or the phone number and
changes the USERNAME_FIELD that is what field will we use as the
username in the login. And this is it.