Created Retrieve, Delete view; added permissions
I created the retrieve view which returns the detail
of a user membership only if is active.
Same with delete view but you just can delete users
from the circle if you are an admin or if you are
the user owner of the membership.
And those are the permissions.